+++ CHANGELOG.md
@@ -4,9 +4,10 @@
 
 ### Breaking changes
 
-* Now supports Qiskit version 0.18.0. As a result of breaking changes
+* Now supports Qiskit version 0.18.3. As a result of breaking changes
   within Qiskit, version 0.17 and below are no longer supported.
   [(#81)](https://github.com/XanaduAI/pennylane-qiskit/pull/81)
+  [(#85)](https://github.com/XanaduAI/pennylane-qiskit/pull/85)
 
 ### Improvements